added common validator function file for common validations
